Diktat-maven-plugin
### What's done:
* Extracted DiktatBaseMojo
* Introduced mojo for fix
